Add memoization utility function
Introduce a new utility function `memoize()` in `memoize`. This function creates a memoized version of a given function by caching its results. It uses a `Map` to store cached results, identified by the JSON stringified version of the arguments. If a cached result is available for the given arguments, it returns that instead of recalculating. This can improve performance for functions with expensive computations and predictable inputs.